Moved test/_utils to testing module. Created a single module for the functions that depend on casacore in msv2_io.py. Updated conftest.py and __init__.py. Updated test_conversion with new testing module. Consolidated documentation from both test and testing README.md.
