+++ tests/unit/measurement_set/_utils/_msv2/test_conversion.py
@@ -10,7 +10,7 @@ import xradio.measurement_set._utils._msv2.conversion as conversion
 from xradio.measurement_set.schema import VisibilityXds
 from xradio.schema.check import check_dataset, check_datatree
 
-from xradio.testing.measurement_set import check_msv4_matches_descr
+from xradio.testing.measurement_set.checker import check_msv4_matches_descr
 
 
 minxds = namedtuple("minxds", "data_vars coords sizes")
